+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Release Notes
 
-## Release 0.15.0
+## Release 0.16.0 (2022-11-18)
+
+### Added
+
+- Automatic replace of `@t-raw` to `@t-out` in `.xml` files during forwardport
+- Added separate command for automatic module migrate `odoo-helper ci auto-migrate-modules`
+  that will try to automatically migrate module code to current odoo version.
+  Currently it can do only simple *find-and-replace* via regex.
+- Added ability to run migration tests with single command, by adding new
+  options to `odoo-helper test` command:
+    - `--migration` - run migration tests
+    - `--migration-repo` - specify repo to test
+
+### Fixes
+
+- support for Odoo 16.0
+
+
+---
+
+## Release 0.15.0 (2022-10-25)
 
 ### Added